pyproject.toml
[project]
name = "kube-claude-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "claude-agent-sdk>=0.1.20",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.1",
]
